Model evaluation: Calculate overall average
Calculate average stats across all SDGs and weight each SDG equally.
Previously, we were calculating the accuracy and other stats across all
texts, which was giving SDGs with more texts a greater weight. Now, each
SDG is given the exact same weight for the calculation of the average,
regardless of the number of associated texts.