Test: Benchmark: Avoid breaking test when expanding benchmark
Using a snapshot test for testing the Benchmark class will cause a
mismatch whenever we expand the benchmark in any way (new SDG,
new texts, etc.). To avoid this, we now test the benchmark output in a
more reliable way that will not need to change when we update the
dataset.